requirements.txt
python-telegram-bot==13.7
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.6.3
tornado>=6.1
